split error and event
